+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ changelog
 * add switch --debug for having the log on console with DEBUG verbosity without need to change configuration file
 * fix regexp for foreign keys when omitting CONSTRAINT keyword
 * change lag display in show_status from seconds to interval for better display
+* add quote_col method in sql_token class to ensure all index columns are correctly quoted
 
 1.2 - 25 May 2017
 ...........................................